% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_formatters.R
\name{format_fa_list}
\alias{format_fa_list}
\title{format_fa_list}
\usage{
format_fa_list(.x, .bullet_col = NULL, .pad_left = "1.75em")
}
\arguments{
\item{.x}{a named list}

\item{.bullet_col}{...}

\item{.pad_left}{...}
}
\value{
a HTML string
}
\description{
format_fa_list
}
\examples{
NULL
}
