% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_button.R
\name{nd_button_block}
\alias{nd_button_block}
\title{nd_button_block}
\usage{
nd_button_block(
  .id,
  .label,
  .fa_class,
  .label_busy = .label,
  .fa_class_busy = .fa_class,
  ...
)
}
\arguments{
\item{.id}{...}

\item{.label}{...}

\item{.fa_class}{...}

\item{.label_busy}{...}

\item{.fa_class_busy}{...}

\item{...}{...}
}
\value{
...
}
\description{
nd_button_block
}
\examples{
NULL
}
