% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_page.R
\name{nd_iframe_app}
\alias{nd_iframe_app}
\title{Title}
\usage{
nd_iframe_app(.url, .width = "100\%", .height = "400pt")
}
\arguments{
\item{.url}{NULL}

\item{.width}{NULL}

\item{.height}{NULL}
}
\description{
Title
}
\examples{
NULL
}
