% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_main.R
\name{nd_main_sec}
\alias{nd_main_sec}
\title{Title}
\usage{
nd_main_sec(..., .bg_color = NULL, .is_grid = TRUE)
}
\arguments{
\item{...}{NULL}

\item{.bg_color}{NULL}

\item{.is_grid}{NULL}
}
\description{
Title
}
\examples{
NULL
}
