% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_page.R
\name{nd_page}
\alias{nd_page}
\title{nd_page}
\usage{
nd_page(
  ...,
  .page_type = "static",
  .navbar = nd_navbar(),
  .main = nd_main(),
  .load_wordcloud2 = TRUE
)
}
\arguments{
\item{...}{NULL}

\item{.page_type}{NULL}

\item{.navbar}{NULL}

\item{.main}{NULL}

\item{.load_wordcloud2}{NULL}
}
\description{
nd_page
}
\examples{
NULL
}
