% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_popover.R
\name{nd_popover}
\alias{nd_popover}
\title{nd_popover}
\usage{
nd_popover(..., .title = "Mehr Informationen", .id = NULL)
}
\arguments{
\item{...}{...}

\item{.title}{...}

\item{.id}{...}
}
\value{
...
}
\description{
nd_popover
}
\examples{
NULL
}
