% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_formatters.R
\name{rlist_c}
\alias{rlist_c}
\title{rlist_c}
\usage{
rlist_c(.x, .fn_leaf_names = NULL, .fn_leaf_wrap = NULL, .fn_node_wrap = NULL)
}
\arguments{
\item{.x}{NULL}

\item{.fn_leaf_names}{NULL}

\item{.fn_leaf_wrap}{NULL}

\item{.fn_node_wrap}{NULL}
}
\value{
string
}
\description{
rlist_c
}
\examples{
NULL
}
